requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
google-generativeai>=0.3.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
comm==0.1.4
confection==0.1.4
contourpy==1.1.1
cookiecutter==2.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
exceptiongroup==1.1.3
executing==1.2.0
fastai==2.7.14
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.18.1
fastprogress==1.0.3
ffmpeg==1.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
fuzzywuzzy==0.18.0
gast==0.5.4
geojson==2.5.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meteostat==1.6.7
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mysql-connector==2.2.9
mysql-connector-python==8.2.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook_shim==0.2.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
openai==1.11.0
openai-whisper==20231117
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
ply==3.11
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.16.1
PyMySQL==1.1.0
Pyomo==6.6.2
pyowm==3.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.39.2
setproctitle==1.3.3
shap==0.43.0
six==1.16.0
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.37.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.7
validators==0.22.0
virtualenv==20.24.5
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.2
yarl==1.9.4
zipp==3.17.0